@@ -7,18 +7,18 @@
         "simple",
         "mvc"
     ],
-    "license": "Apache-2.0",
+    "license": "MIT",
     "require": {
-        "php": ">=7.3",
+        "php": "^7.4 || ^8.0",
         "league/plates": "^3.3",
         "php-di/php-di": "^6.0",
         "nikic/fast-route": "^1.3",
-        "nyholm/psr7": "^1.4",
+        "nyholm/psr7": "^1.5",
         "nyholm/psr7-server": "^1.0",
         "psr/log": "^1.1"
     },
     "require-dev": {
-        "phpstan/phpstan": "^0.12",
+        "phpstan/phpstan": "^1.7",
         "phpunit/phpunit": "^9.5"
     },
     "autoload": {
